% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionsEnrichment.R
\name{get_score}
\alias{get_score}
\title{Create score depending on threshold and paradigm}
\usage{
get_score(genes, fc, thr.fc, fdr, thr.fdr, ntop)
}
\arguments{
\item{genes}{character vector with list of ENSEBML IDs}

\item{fc}{numeric vector of logFC for each gene}

\item{thr.fc}{significance threshold for absolute logFC}

\item{fdr}{numeric vector of FDR for each gene}

\item{thr.fdr}{significance threshold for FDR}

\item{ntop}{number of first taken genes}
}
\value{
numeric score vector
}
\description{
Create score depending on threshold and paradigm
}
