% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_X_num}
\alias{get_X_num}
\title{Convert input object as numeric matrix}
\usage{
get_X_num(obj, assay_string = NULL)
}
\arguments{
\item{obj}{input data with features in rows and samples in columns. 
Could be a `SummarizedExperiment` object, matrix or `Seurat` object. 
For `SummarizedExperiment` with multiple assays or `Seurat` pass the name 
with `assay_string` parameter, otherwise the first will be taken.
See \code{\link[SummarizedExperiment]{SummarizedExperiment-class}}}

\item{assay_string}{name of assay for `SummarizedExperiment` or `Seurat` 
input object `obj`. Default value is NULL}
}
\value{
matrix
}
\description{
Convert input object as numeric matrix
}
\keyword{internal}
