% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coreICA.R
\name{outICA}
\alias{outICA}
\title{Outside part of multiple run Independent Component Analysis}
\usage{
outICA(X, n.comp, row.norm = FALSE, verbose = FALSE)
}
\arguments{
\item{X}{matrix with features in rows and samples in columns}

\item{n.comp}{number of components}

\item{row.norm}{rows normalization flag. Default value is FALSE}

\item{verbose}{logic TRUE or FALSE. Use TRUE for print process steps. 
Default value is FALSE}
}
\value{
a list with 
        \item{X}{input matrix}
        \item{X1}{interim calculated matrix}
        \item{K}{pre-whitening matrix that projects data onto the first 
       `n.comp` principal components}
}
\description{
Calculate a common part for consensus Independent Component 
Analysis (ICA)
}
\author{
Maryna Chepeleva
}
