% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sortDataFrame}
\alias{sortDataFrame}
\title{Sort dataframe}
\usage{
sortDataFrame(x,key, ...)
}
\arguments{
\item{x}{a data.frame}

\item{key}{sort by this column}

\item{...}{other parameters for `order` function (e. g. `decreasing`)}
}
\value{
sorted dataframe
}
\description{
Sort dataframe, adapted from  
http://snippets.dzone.com/user/r-fanatic
}
\examples{
df <- data.frame("features" = c("f1", "f2", "f3"), fdr = c(0.02, 0.002, 1))
sortDataFrame(df, "fdr")
}
