% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sortFeatures}
\alias{sortFeatures}
\title{Sort Genes of consICA object}
\usage{
sortFeatures(Genes)
}
\arguments{
\item{Genes}{list compilant to `getFeatures`output}
}
\value{
sorted list
}
\description{
Sort Genes for independent components
}
\examples{
#features <- list("ic1" = list(
#              "pos" = data.frame("features" = c("f1", "f2", "f3"),
#                                 "fdr" = c(0.0043, 0.4, 0.04)), 
#              "neg" = data.frame("features" = c("f1", "f2", "f3"),
#                                 "fdr" = c(0, 0.1, 0.9))))
#sortFeatures(features)
}
\keyword{internal}
