% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGO.R
\name{getGO}
\alias{getGO}
\title{Assigns IC signatures to Gene Ontologies}
\usage{
getGO(
  IC,
  alpha = 0.05,
  genenames = NULL,
  genome = "org.Hs.eg.db",
  db = c("BP", "CC", "MF")
)
}
\arguments{
\item{IC}{list compliant to `consICA()` result}

\item{alpha}{value in [0,1] interval. Used to filter features with 
FDR < `alpha`. Default value is 0.05}

\item{genenames}{alternative names of genes}

\item{genome}{R-package for genome annotation used. For human -
'org.Hs.eg.db'}

\item{db}{name of GO database: "BP","MF","CC"}
}
\value{
list for each db chosen (BP, CC, MM), with dataframes `pos` for 
positive and `neg` for negative affecting features for each component:
    \item{GO.ID}{id of gene ontology term}
    \item{Term}{name of term}
    \item{Annotated}{number of annotated genes}
    \item{Significant}{number of significant genes}
    \item{Expected}{estimate of the number of annotated genes if the 
    significant} 
    \item{genes would be randomly selected from the gene universe
    classisFisher}{F-test}
    \item{FDR}{false discovery rate value}
    \item{Score}{genes score}
}
\description{
Assigns extracted independent components to Gene Ontologies
}
\examples{
data("samples_data")
cica <- consICA(samples_data$X, ncomp=2, ntry=1, show.every=0)
GOs <- getGO(cica, db = "BP")
}
\author{
Petr V. Nazarov
}
