% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consICA.R
\name{oneICA}
\alias{oneICA}
\title{Runs \code{\link[fastICA]{fastICA}}}
\usage{
oneICA(
  X,
  ncomp = 10,
  filter.thr = NULL,
  reduced = FALSE,
  fun = "logcosh",
  alg.typ = "deflation"
)
}
\arguments{
\item{X}{matrix with features in rows and samples in columns}

\item{ncomp}{number of components. Default value is 10}

\item{filter.thr}{filter rows in input matrix with max value > `filter.thr`.
Default value is NULL}

\item{reduced}{If TRUE returns reduced result (no X, see 'return')}

\item{fun}{the functional form of the G function used in the approximation 
to neg-entropy in fastICA. Default value is "logcosh"}

\item{alg.typ}{parameter for fastICA(). if alg.typ == "deflation" the 
components are extracted one 
at a time. if alg.typ == "parallel" the components are extracted 
simultaneously. Default value is "deflation"}
}
\value{
a list with
        X - data
        nsples, nfeatures - dimentions of X
        S, M - consensus metagene and weight matrix
        ncomp - number of components
}
\description{
Runs \code{\link[fastICA]{fastICA}} once and store in a 
consICA manner
}
\examples{
data("samples_data")
res <- oneICA(samples_data$X)
}
\seealso{
\code{\link[fastICA]{fastICA}}
}
\author{
Petr V. Nazarov
}
