% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{samples_data}
\alias{samples_data}
\title{Samples of gene expression}
\format{
A SummarizedExperiment object:
\describe{
  \item{assay}{expression matrix with genes by rows and samples by columns}
  \item{colData}{data frame with sample metadata (clinical variables)}
}
}
\usage{
data(samples_data)
}
\description{
A dataset containing the expression of 2454 genes for 472 samples from skin 
cutaneous melanoma (SKCM) TCGA cohort, their metadata such as age, gender, 
cancer type etc. and survival time-to-event data
}
\keyword{datasets}
