% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_bpparam}
\alias{set_bpparam}
\title{Set up for the parallel computing for biocParallel
Adapt from `FEAST`
This function sets up the environment for parallel computing.}
\usage{
set_bpparam(ncores = 0, BPPARAM = NULL)
}
\arguments{
\item{ncores}{number of processors}

\item{BPPARAM}{bpparameter from bpparam}
}
\value{
BAPPARAM settings
}
\description{
Set up for the parallel computing for biocParallel
Adapt from `FEAST`
This function sets up the environment for parallel computing.
}
\keyword{internal}
